Move gateway to own directory and rename
Renamed app to gateway and moved all files to new dir. Changed CI to work properly.sign test
