+++ gatewayApp/pytest.ini
@@ -2,3 +2,4 @@
 filterwarnings = 
     ignore::DeprecationWarning:websockets.connection
     ignore:pkg_resources is deprecated:DeprecationWarning
+    ignore:Deprecated call to `pkg_resources.declare_namespace`
